Created more folders to organize components
