+++ ecosphere-goods/src/layouts/MainLayout.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import Navbar from '../components/Navbar'
-import Footer from '../components/Footer'
+import Navbar from '../components/Layout/Navbar'
+import Footer from '../components/Layout/Footer'
 import { Outlet } from 'react-router-dom'
 
 const MainLayout = ({ children }) => {
