+++ ecosphere-goods/src/pages/LoginPage.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const LoginPage = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default LoginPage
